More playing with owen scrambling.
Ultimately, it actually looks like the v2 improvement is an even
better hash then I thought.  The v3 hash is basically the same
construction, but allowing more rounds with additional random numbers,
and it has the nice the property that the more rounds you throw at it,
the better the mixing and the less and less bias in the avalanche.

The LK hash, by contrast, never reduces its bias... which is really bad.
It seems the *only* reason it works reasonably is because of adding the
scramble value at the beginning.  But the actual rounds are shockingly
ineffective.